requirements.txt
huawei_lte_api==1.9.3
speedtest-cli==2.1.3
